% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_signature_genes.R
\name{calculate_snr}
\alias{calculate_snr}
\title{Function to calculate Signal-to-noise Ratio between IN and OUT populations}
\usage{
calculate_snr(gem, true_cells, false_cells)
}
\arguments{
\item{gem}{Data frame of signature genes in cells}

\item{true_cells}{list of cell names that are confidently IN-population
should match with column names of gem}

\item{false_cells}{list of cell names that are confidently OUT-population
should match with column names of gem}
}
\value{
list of SNR value for each signature gene
}
\description{
Function to calculate Signal-to-noise Ratio between IN and OUT populations
}
