% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_signature_genes.R
\name{choose_training_set}
\alias{choose_training_set}
\title{Function to choose training IN and OUT populations using precision-recall}
\usage{
choose_training_set(gem, positive_markers, negative_markers)
}
\arguments{
\item{gem}{Data frame of gene expression of genes (rows) in cells (columns)}

\item{positive_markers}{List of gene names expected to be upregulated in IN population}

\item{negative_markers}{List of gene names expected to be downregulated in IN population}
}
\value{
Lists of training IN and OUT cells
}
\description{
Function to choose training IN and OUT populations using precision-recall
}
