% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_signature_genes.R
\name{choose_unsupervised}
\alias{choose_unsupervised}
\title{Function to choose initial IN and OUT populations with linear regression}
\usage{
choose_unsupervised(gem, signature_genes)
}
\arguments{
\item{gem}{Data frame of gene expression of genes (rows) in cells (columns)}

\item{signature_genes}{list of gene names (should be same format as rownames of gem)}
}
\value{
a lists of confident IN and OUT cells
}
\description{
Function to choose initial IN and OUT populations with linear regression
}
