% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_signature_genes.R
\name{choose_with_negative_markers}
\alias{choose_with_negative_markers}
\title{Function to select initial OUT population from negative markers}
\usage{
choose_with_negative_markers(gem)
}
\arguments{
\item{gem}{Data frame of expression of negative markers (rows) in cells of the dataset (columns)}
}
\value{
list of confident OUT population
}
\description{
Function to select initial OUT population from negative markers
}
