% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_signature_genes.R
\name{choose_with_positive_markers}
\alias{choose_with_positive_markers}
\title{Function to select initial IN population from necessary positive markers}
\usage{
choose_with_positive_markers(gem)
}
\arguments{
\item{gem}{Data frame of expression of positive/necessary markers (rows) in cells of 
the dataset (columns)}
}
\value{
lists of confident IN (cells that express all the positive markers)
and confident OUT cells (cells that do not express any of the positive markers)
}
\description{
Function to select initial IN population from necessary positive markers
}
