% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_functions.R
\name{counts_to_cpm}
\alias{counts_to_cpm}
\title{Function to convert gene counts CPM-normalized}
\usage{
counts_to_cpm(counts_gem)
}
\arguments{
\item{gem}{Data frame of}
}
\value{
CPM-normalized gene expression matrix
}
\description{
Function to convert gene counts CPM-normalized
}
