% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropout_correction.R
\name{EM_for_dropout_estimation}
\alias{EM_for_dropout_estimation}
\title{Estimate Exponential and Normal distribution parameters with EM algorithm
from single-cell data}
\usage{
EM_for_dropout_estimation(gem)
}
\arguments{
\item{gem}{Gene expression matrix, rows are genes and columns are cells
gem should be normalized and log-transformed (log2(gem + 1.01))
1.01 used instead of 1 to avoid zero values}
}
\value{
a table with the estimated dropout probabilities for each gene in each cell of the gem

a list of estimated parameters for each gene
}
\description{
Estimate Exponential and Normal distribution parameters with EM algorithm
from single-cell data
}
