% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_GMM.R
\name{final_populations}
\alias{final_populations}
\title{Function to identify matching population from scID score by fitting a Gaussian finite mixture model from Mclust}
\usage{
final_populations(score)
}
\arguments{
\item{score}{List of scID scores for target cells}
}
\value{
List of matching cells
}
\description{
Function to identify matching population from scID score by fitting a Gaussian finite mixture model from Mclust
}
