% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markers.R
\name{find_markers}
\alias{find_markers}
\title{Function to extract cluster specific genes from reference clusters.
This function uses MAST test as implemented in the Seurat package.}
\usage{
find_markers(
  reference_gem,
  reference_clusters,
  logFC,
  only.pos,
  normalize_reference
)
}
\arguments{
\item{reference_gem}{Data frame of gene expression (rows) per cell (columns) in reference data}

\item{reference_clusters}{Named list of cluster IDs of reference cells}

\item{logFC}{LogFC threshold for extracting markers from reference clusters}

\item{only.pos}{Logical to include negative markers in the cluster specific gene sets}

\item{normalize_reference}{Logical to select if reference data need to be normalized (when raw counts have been provided)}
}
\value{
Data frame of cluster specific genes extracted
}
\description{
Function to extract cluster specific genes from reference clusters.
This function uses MAST test as implemented in the Seurat package.
}
