% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{gmt_to_markers}
\alias{gmt_to_markers}
\title{Function to read markers from .gmt file format (Genomic Cytometry)}
\usage{
gmt_to_markers(gmt_file)
}
\arguments{
\item{filename}{.gmt file with markers}
}
\value{
Data frame of cluster specific genes per celltype
}
\description{
Function to read markers from .gmt file format (Genomic Cytometry)
}
