% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropout_correction.R
\name{initialize_prmt_exp}
\alias{initialize_prmt_exp}
\title{Initialize vector of paramaters for the mixture model}
\usage{
initialize_prmt_exp(x)
}
\arguments{
\item{x}{vectors of a gene's expression across all cells}
}
\value{
a named vector with the initial parameters: [lambda, theta, mu, sigma]
}
\description{
Initialize vector of paramaters for the mixture model
}
