% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{loadfast}
\alias{loadfast}
\title{Function to read Gene Expression Matrix from file}
\usage{
loadfast(filename, header = T)
}
\arguments{
\item{filename}{directory and name of input file}

\item{header}{specify if first line of file is the field names}
}
\value{
Data frame of GEM with genes in rows and cells in columns
}
\description{
Function to read Gene Expression Matrix from file
}
