% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_functions.R
\name{normalize_gene}
\alias{normalize_gene}
\title{Function to row-normalize a numerical vector by the 99th percentile}
\usage{
normalize_gene(x)
}
\arguments{
\item{x}{Numerical vector}
}
\value{
Vector with normalized values
}
\description{
Function to row-normalize a numerical vector by the 99th percentile
}
