% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_functions.R
\name{plot_score_2D}
\alias{plot_score_2D}
\title{Function to plot heatmap of average cluster-specific geneset expression in clusters of cells}
\usage{
plot_score_2D(gem, labels, markers, clusterID, weights)
}
\arguments{
\item{gem}{Data frame of gene expression (rows) per cell (columns) in target data}

\item{labels}{List of cluster IDs for cells of gem}

\item{markers}{Data frame of cluster specific genes with at least "gene", "cluster" and "avg_logFC" columns}

\item{clusterID}{cluster ID of cluster of interest}

\item{weights}{list of weighst of cluster specific genes as returned by scid_multiclass}
}
\description{
Function to plot heatmap of average cluster-specific geneset expression in clusters of cells
}
