% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scID.R
\name{scid_multiclass}
\alias{scid_multiclass}
\title{Main function to get Gene expression matrix and signature genes and return matches and scores}
\usage{
scid_multiclass(target_gem = NULL, reference_gem = NULL,
  reference_clusters = NULL, markers = NULL, logFC = 0.5,
  estimate_weights_from_target = FALSE, weights = NULL,
  only_pos = FALSE, normalize_reference = TRUE)
}
\arguments{
\item{target_gem}{Data frame of gene expression (rows) per cell (columns) in target data}

\item{reference_gem}{Data frame of gene expression (rows) per cell (columns) in reference data}

\item{reference_clusters}{Named list of cluster IDs of reference cells}

\item{markers}{Data frame of cluster specific genes that will be used instead of the reference data}

\item{logFC}{LogFC threshold for extracting markers from reference clusters}

\item{only_pos}{Logical to include negative markers in the cluster specific gene sets}

\item{normalize_reference}{Logical to select if reference data need to be normalized (when raw counts have been provided)}

\item{use_reference_for_weights}{Logical to use either reference or target data for sorting the signature genes}
}
\value{
list of cluster IDs for target cells

scID scores for target cells

markers used

estimated gene weights
}
\description{
Main function to get Gene expression matrix and signature genes and return matches and scores
}
