% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scID.R
\name{scid_singleclass}
\alias{scid_singleclass}
\title{Main function to get Gene expression matrix and signature genes and return matches and scores}
\usage{
scid_singleclass(target_gem = NULL, gene_signature = NULL,
  likelihood_threshold = 0.95)
}
\arguments{
\item{target_gem}{Data frame of gene expression (rows) per cell (columns) in target data}

\item{gene_signature}{List of genes symbols enriched in the population of interest}

\item{likelihood_threshold}{Minimum required likelihood of gene signature score for a cell to be assigned to the respective reference cluster}
}
\value{
list of target cells matching to the population of interest

list of matching scores for all target cells
}
\description{
Main function to get Gene expression matrix and signature genes and return matches and scores
}
