% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_signature_genes.R
\name{scID_weight}
\alias{scID_weight}
\title{Main function for estimation of gene ranks}
\usage{
scID_weight(gem, true_cells, false_cells)
}
\arguments{
\item{gem}{Data frame of signature genes in cells}

\item{true_cells}{list of cell names that are confidently IN-population
should match with column names of gem}

\item{false_cells}{list of cell names that are confidently OUT-population
should match with column names of gem}
}
\value{
list of weights for each signature gene
}
\description{
Main function for estimation of gene ranks
}
