% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropout_correction.R
\name{update_prmt_exp}
\alias{update_prmt_exp}
\title{Update parameters}
\usage{
update_prmt_exp(d, x)
}
\arguments{
\item{d}{The posterior latent distribution}

\item{x}{The vector of gene expression across all cells}
}
\value{
the parameters lambda, theta, mu and sigma
}
\description{
Update parameters
}
