% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_signature_genes.R
\name{weight_signature}
\alias{weight_signature}
\title{Main function for estimation of gene ranks}
\usage{
weight_signature(gem, signature, positive_markers = NULL,
  negative_markers = NULL)
}
\arguments{
\item{gem}{Data frame of gene expression (rows) in cells (columns)}

\item{signature}{List of signature gene names (should match with rownames of gem)}

\item{positive_markers}{(optional) List of genes that should be present in all confident IN-cells}

\item{negative_markers}{(optional) List of genes that should not be present in any IN-cell}
}
\value{
List of weights per gene
}
\description{
Main function for estimation of gene ranks
}
