% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scID.R
\name{scid_match_cells}
\alias{scid_match_cells}
\title{Main function to get Gene expression matrix and signature genes and return matches and scores}
\usage{
scid_match_cells(signature_file = NULL, gem_file = NULL, scData = NULL,
  signature_genes = NULL, contamination = 0, sort.signature = TRUE,
  gene.weights = NULL)
}
\arguments{
\item{signature_file}{(optional) filename of the signature genes}

\item{gem_file}{(optional) filename of the gene expression matrix}

\item{scData}{(optional) Data frame of gene expression (rows) per cell (columns)
either gem_file or scData must be given as input}

\item{signature_genes}{(optional) a list of the signature genes
either signature_file or signature_genes must be given as input}

\item{contamination}{Percentage of accepted cells that belong to the common area between 
IN and OUT population distributions}

\item{do.imputation}{Logical to choose not to correct for dropouts (default is TRUE for doing imputation)}

\item{gene.weigts}{list of weights for the signature genes if known}
}
\value{
list of names of matching cells (IN-population) and

matching score of every cell of the dataset

list of weight per gene
}
\description{
Main function to get Gene expression matrix and signature genes and return matches and scores
}
