% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scID.R
\name{scid_match_cells}
\alias{scid_match_cells}
\title{Main function to get Gene expression matrix and signature genes and return matches and scores}
\usage{
scid_match_cells(target_gem = NULL, reference_gem = NULL,
  reference_clusters = NULL, logFC = 0.5,
  use_reference_for_weights = FALSE, likelihood_threshold = 0.99)
}
\arguments{
\item{target_gem}{Data frame of gene expression (rows) per cell (columns) in target data}

\item{reference_gem}{Data frame of gene expression (rows) per cell (columns) in reference data}

\item{reference_clusters}{Named list of cluster IDs of reference cells}

\item{logFC}{LogFC threshold for extracting markers from reference clusters}

\item{use_reference_for_weights}{Logical to use either reference or target data for sorting the signature genes}

\item{likelihood_threshold}{Minimum required likelihood of gene signature score for a cell to be assigned to the respective reference cluster}
}
\value{
list of target cluster IDs
}
\description{
Main function to get Gene expression matrix and signature genes and return matches and scores
}
