% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_signature_genes.R
\name{scID_weight}
\alias{scID_weight}
\title{Main function for estimation of gene ranks}
\usage{
scID_weight(gem, true_cells, false_cells)
}
\arguments{
\item{gem}{Data frame of signature genes in cells}

\item{true_cells}{List of training IN cells}

\item{false_cells}{List of training OUT cells}
}
\value{
List of weights for signature genes
}
\description{
Main function for estimation of gene ranks
}
